refactor: make package and rename `route` to `graph`
